client/requirements.txt
prompt_toolkit==3.0.24
mysql_connector_repackaged==0.3.1
prettytable==2.4.0
